feat!: replace `EncoderLocalizationDemo` with `KiwiDriveLocalizationDemo`
